Creación de la ventana de alquileres. Funciona teniendo en cuenta la base de datos!!...

+++ videoclub/src/com/EnjoyVideoClub/Model/Pelicula.java
@@ -25,7 +25,7 @@ public class Pelicula extends Multimedia{
         setActrizPrincipal(actrizPrincipal);
     }
 
-    public double getDuracionPelicula() {
+    public int getDuracionPelicula() {
         return duracionPelicula;
     }
     public void setDuracionPelicula(int duracionPelicula) {
